requirements.txt
Flask==3.0.3
tensorflow-cpu==2.16.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.2
Pillow==10.3.0
Werkzeug==3.0.1
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.2
gunicorn==22.0.0
python-dotenv==1.0.0
psutil==5.9.8
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
bidict==0.22.1
efficientnet==1.1.1